refactor: move decorative svg images to own folder 'decorative-bg'
